solver: move Result and the solver errors out of asp.py
Split is the following:
* spack.solver.result carries Result, the optimization criteria types
  and the SpecDict serializers.
* spack.solver.error carries the exceptions and format_unsolved.

Neither imports the solver, so a caller can be typed against a result
without pulling in how it is produced.

No behavior changed.
